requirements.txt
fastapi[all]
google
google-genai
elevenlabs
python-dotenv
uvicorn
requests